Flight AI 8.0/backend/requirements.txt
# Backend requirements
flask
openai
sqlalchemy

flask-cors
gunicorn
pytest
python-dotenv
flask-limiter
flask-talisman
redis>=4.0
